Lab5/consumer1/requirements.txt
kafka-python==2.2.2